requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
amqp==1.4.6               # via kombu
anyjson==0.3.3            # via kombu
billiard==3.3.0.20        # via celery
celery==3.1.18
cssselect==0.9.1          # via pyquery
django-grappelli==2.7.1
django==1.8.3
djangorestframework==3.2.2
ecdsa==0.13               # via paramiko
gunicorn==19.3.0
kombu==3.0.26             # via celery
lxml==3.4.4               # via pyquery
paramiko==1.15.2
pillow==2.9.0
pycrypto==2.6.1           # via paramiko
pyquery==1.2.9
pytz==2015.4              # via celery
